pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "review-iq"
version = "0.1.0"
description = "Production-grade review intelligence service. Unstructured customer reviews → queryable structured insights."
authors = [{ name = "Gaurav Gandhi", email = "gaurav.gandhi2411@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "groq>=0.13.0",
    "structlog>=25.1.0",
    "prometheus-client>=0.21.0",
    "slowapi>=0.1.9",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.19",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "google-genai>=1.0.0",
    "supabase>=2.10.0",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "argon2-cffi>=25.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "types-requests>=2.32.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app", "tests", "eval"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = false
exclude = ["tests/", "eval/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term-missing -v --cov-fail-under=60 -m 'not integration'"
markers = [
    "integration: tests requiring a live Supabase/Postgres connection (run with -m integration)",
]

[tool.coverage.run]
omit = [
    "app/templates/*",
    "app/static/*",
]
